pyproject.toml
[project]
name = "proj-eng2025"
version = "0.1.0"
description = ""
authors = [
    {name = "Felipe Alves",email = "lipealvessax@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kafka-python (==2.0.2)",
    "pandas (==2.2.2)",
    "requests (==2.31.0)",
    "python-dotenv (==1.1.1)",
    "streamlit (==1.35.0)",
    "faker (==25.2.0)",
    "pymongo (==4.7.2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
